fix token issue for Claude
